refactor: improve redux structure
