--- src/app/features/lp-optimisers/helpers.ts
@@ -1,8 +1,0 @@
-export const rejectThunkWithError = (
-  thunkAPI: {
-    rejectWithValue: (value: string | undefined) => unknown;
-  },
-  err: unknown,
-) => {
-  return thunkAPI.rejectWithValue((err as Error)?.message);
-};
